requirements.txt
Pillow>=9.0.0
exifread>=2.3.0
tkinterdnd2>=0.3.0




